fix: divide code into more modules, rename some modules and functions, correct mistakes
