+++ index.html
@@ -235,8 +235,8 @@
     </section>
   </template>
 
-  <script src="js/main.js" type="module"></script>
   <script src="vendor/pristine/pristine.min.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
