+++ js/render-gallery.js
@@ -1,5 +1,8 @@
-import { photos } from './create-photos-array.js';
+import { createPhotosArray } from './create-photos-array.js';
+import { clickPhotoHandler } from './click-photo-handler.js';
+import { uploadPhotoHandler } from './upload-photo-handler.js';
 
+const photos = createPhotosArray();
 const photosContainer = document.querySelector('.pictures');
 const photoTemplate = document.querySelector('#picture').content.querySelector('.picture');
 const photosListFragment = document.createDocumentFragment();
@@ -14,11 +17,11 @@ const createListFragment = ({id, url, description, likes, comments}) => {
   photosListFragment.append(newPhoto);
 };
 
-const renderPhotos = (photosArray) => {
-  photosArray.forEach(createListFragment);
+const renderGallery = () => {
+  photos.forEach(createListFragment);
   photosContainer.append(photosListFragment);
+  clickPhotoHandler();
+  uploadPhotoHandler();
 };
 
-renderPhotos(photos);
-
-export { photos, photosContainer };
+export { photos, photosContainer, renderGallery };
